Move Scope.element to Cursor.element
We have to remember the current element being parsed but that is
independent from the current scope.
But we have to start parsing a new element without already starting a new
scope, e.g. when we want to modify the scope through one element.

So remember the current element within our Cursor.